+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master
 
 * Fix padding, background, and color contrast in `NumberedCodeSnippet` component. [#269](https://github.com/mapbox/dr-ui/pull/269)
+* Update mr-ui and other dependencies for security fixes. [#266](https://github.com/mapbox/dr-ui/pull/266)
 
 ## 0.27.0